Suppr lignes inutiles, commentaires et documentation

+++ toudou_starter_models/toudou/src/toudou/models.py
@@ -1,9 +1,10 @@
 import csv
 from dataclasses import dataclass
 from datetime import datetime
-from typing import Optional
+from typing import Optional, List
 import sqlalchemy as db
-from sqlalchemy import create_engine, MetaData, Table, Column, Integer, String, Boolean, DateTime
+from sqlalchemy import create_engine, MetaData, Table, Column, Integer, String, Boolean, DateTime, Row
+from sqlalchemy.engine.result import _TP
 
 TODO_FOLDER = "db"
 metadata_obj = MetaData()
@@ -29,14 +30,13 @@ class Todo:
 def init_db() -> None:
     metadata_obj.create_all(engine)
 
-def write_to_bd(todo: Todo) -> None:
+
+def write_to_bd(todo: Todo) -> bool:
+    """
+    Enregistre une tâche [todo] à la base de données si son id n'y est pas.
+    """
     if not get_todo(todo.id):
-        requete = db.Insert(todo_table).values(
-            id=todo.id,
-            task=todo.task,
-            complete=todo.complete,
-            due=todo.due
-        )
+        requete = db.Insert(todo_table).values(id=todo.id, task=todo.task, complete=todo.complete, due=todo.due)
         with engine.connect() as conn:
             conn.execute(requete)
             conn.commit()
@@ -45,21 +45,21 @@ def write_to_bd(todo: Todo) -> None:
 
 
 def update_bd(todo: Todo) -> None:
-    requete = db.Update(todo_table).values(
-        id=todo.id,
-        task=todo.task,
-        complete=todo.complete,
-        due=todo.due
-    )
-    col_id = db.sql.column('id')
-    requete = requete.where(col_id == todo.id)
+    """
+    Modifie une tâche dans la base de données avec son id
+    """
+    requete = db.Update(todo_table).values(id=todo.id, task=todo.task, complete=todo.complete, due=todo.due)
+    requete = requete.where(db.sql.column('id') == todo.id)
     with engine.connect() as conn:
         conn.execute(requete)
         conn.commit()
 
 
-def get_next_id():
-    # devieng obsoloete si 10000000 de taches
+def get_next_id() -> int:
+    """
+    Génere un nouvel id selon ceux des tâches deja enregistrés (max des id +1).
+    Limite: Si le nombre de tâches devient trop grand.
+    """
     with engine.connect() as conn:
         result = conn.execute(db.text("""SELECT MAX(id) FROM todos""")).fetchone()
         conn.commit()
@@ -69,12 +69,18 @@ def get_next_id():
 
 
 def create_todo(task: str, complete: bool = False, due: Optional[datetime] = None) -> None:
+    """
+    Crée une tâche.
+    """
     id = get_next_id()
     todo = Todo(id=id, task=task, complete=complete, due=due)
     write_to_bd(todo)
 
 
 def get_todo(par_id: int) -> Todo | bool:
+    """
+    Retourne la tâche (objet Todo) correspondant à l'id (par_id).
+    """
     col_id = db.sql.column('id')
     requete = todo_table.select()
     requete = requete.where(col_id == par_id)
@@ -88,63 +94,63 @@ def get_todo(par_id: int) -> Todo | bool:
 
 
 def get_todos() -> list:
+    """
+    Retourne une liste de tâches triées par date sous forme de tuple.
+    Ex : [(id,task,complete,due),...]
+    """
     requete = todo_table.select()
     with engine.connect() as conn:
         result = conn.execute(requete).fetchall()
         conn.commit()
-    return sorted(result,key=lambda x : x[3])
-
+    return sorted(result, key=lambda x: x[3])
     # Autre methode :
     # requete = select(todo_table)
     # result = engine.connect().execute(requete).fetchall()
 
 
-def update_todo(
-        id: int,
-        task: str,
-        complete: bool,
-        due: Optional[datetime]
-) -> None:
+def update_todo(id: int, task: str, complete: bool, due: Optional[datetime]) -> None:
+    """
+    Modifie une tâche (l'objet Todo) en fonction des paramêtres et l'enregistre dans la bd.
+    """
     if get_todo(id):
         todo = get_todo(id)
         todo.task = task
         todo.complete = complete
         due = due.split("-")
-        todo.due = datetime(int(due[0]), int(due[1]), int(due[2])) if due else todo.due  # coz can be none
+        todo.due = datetime(int(due[0]), int(due[1]), int(due[2])) if due else todo.due  # pourrait être None
         update_bd(todo)
 
 
-def delete_todo(id: int) -> None:
-    col_id = db.sql.column('id')
-    requete = todo_table.delete()
-    requete = requete.where(col_id == id)
-    with engine.connect() as conn:
-        result = conn.execute(requete)
-        conn.commit()
+def delete_todo(id: int) -> bool:
+    """
+    Supprime une tâche de la bd en fonction de [id]
+    Retourne True en cas de succès, False en cas d'échec
+    """
+    if get_todo(id):
+        col_id = db.sql.column('id')
+        requete = todo_table.delete()
+        requete = requete.where(col_id == id)
+        with engine.connect() as conn:
+            conn.execute(requete)
+            conn.commit()
+        return True
+    return False
 
 
 def complete_todo(par_id: int) -> None:
+    """
+    Change le champs 'complete' d'un objet Todo en True.
+    """
     tache = get_todo(par_id)
     tache.complete = True
     update_bd(tache)
 
 
 def delete_all() -> None:
+    """
+    Supprime toutes les tâches contenues dans la bd.
+    """
     requete = todo_table.delete()
     with engine.connect() as conn:
-        result = conn.execute(requete)
+        conn.execute(requete)
         conn.commit()
-
-
-def import_from_csv(file: str) -> list:
-    i = 0
-    tasks = []
-    with open(file, 'r') as file:
-        csvreader = csv.reader(file, delimiter=',')
-        for r in csvreader:
-            if i > 0:
-                date = [int(e) for e in r[3].split(' ')[0].split('-')]
-                tasks.append(Todo(int(r[0]), r[1], True if r[2] == 'True' else False,
-                                  datetime(date[0], date[1], date[2])))
-            i += 1
-        return tasks
